+++ kotlinpoet/src/jvmTest/kotlin/com/squareup/kotlinpoet/CodeBlockTest.kt
@@ -47,6 +47,20 @@ class CodeBlockTest {
   }
 
   @Test
+  fun addComment() {
+    val codeBlock = CodeBlock.builder().addComment("a comment").build()
+    assertThat(codeBlock.toString()).isEqualTo("// a comment\n")
+  }
+
+  @Test
+  fun addCommentWithPlaceholders() {
+    val codeBlock =
+      CodeBlock.builder().addComment("not yet implemented %L of type %T", "field", STRING).build()
+    assertThat(codeBlock.toString())
+      .isEqualTo("// not yet implemented field of type kotlin.String\n")
+  }
+
+  @Test
   fun doublePrecision() {
     val doubles =
       listOf(